Split starlark.go into multiple files
It was becoming a bit larger than I like.
