+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include go.mod
 include go.sum
+include *.go
 include starlark.c
-include starlark.go
 include starlark.h
